move test logger
